@@ -11,36 +11,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.1.2",
+    "@angular/animations": "~8.2.14",
     "@angular/cdk": "~7.1.1",
-    "@angular/common": "~7.1.2",
-    "@angular/compiler": "~7.1.2",
-    "@angular/core": "~7.1.2",
-    "@angular/forms": "~7.1.2",
-    "@angular/http": "~7.1.2",
+    "@angular/common": "~8.2.14",
+    "@angular/compiler": "~8.2.14",
+    "@angular/core": "~8.2.14",
+    "@angular/forms": "~8.2.14",
     "@angular/material": "^7.1.1",
-    "@angular/platform-browser": "~7.1.2",
-    "@angular/platform-browser-dynamic": "~7.1.2",
-    "@angular/router": "~7.1.2",
-    "@ngrx/effects": "^6.1.2",
-    "@ngrx/store": "^6.1.2",
-    "@ngrx/store-devtools": "^6.1.2",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/router": "~8.2.14",
+    "@ngrx/effects": "^8.5.2",
+    "@ngrx/store": "^8.5.2",
+    "@ngrx/store-devtools": "^8.5.2",
     "angular-in-memory-web-api": "^0.7.0",
     "core-js": "^2.5.4",
-    "rxjs": "~6.3.3",
+    "har-validator": "^5.1.3",
+    "rxjs": "~6.5.3",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "~7.1.2",
-    "@angular/compiler-cli": "~7.1.2",
-    "@angular/language-service": "~7.1.2",
-    "@ngrx/schematics": "^6.1.2",
+    "@angular-devkit/build-angular": "~0.803.20",
+    "@angular/cli": "~8.3.20",
+    "@angular/compiler-cli": "~8.2.14",
+    "@angular/language-service": "~8.2.14",
+    "@ngrx/schematics": "^8.5.2",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.0.0",
@@ -51,6 +51,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.1"
+    "typescript": "~3.5.3"
   }
 }
